fix: PathSegment::Index is u64, and the width rule is exhaustive
`PathSegment::Index` is a zero-based ordinal into a list in a produced result.
It is neither a line or column ordinal nor a byte offset, so neither half of
the rule the last commit wrote down reached it, and it kept tokora's `u32` by
inheritance rather than by decision.

Widened, for the reason that decided the ordinals: the crate is unpublished so
it is free now and breaking later, and the type should not imply a ceiling the
domain does not have. Unlike `Location::source` there is no already-`u32`
contract to round-trip through — a result path entry is "an integer" in the
specification and a number in JSON, so a narrower type would be a cap invented
to justify itself. The adapter converts from tokora's `u32` with `u64::from`,
which loses nothing; painty differing from tokora here is the point rather
than an awkwardness, since painty's model is deliberately not shaped by
tokora's.

THE RULE MATTERED MORE THAN THE FIELD. Two categories were named and the type
system held more, so a member sat unplaced and no test could notice — an
ascription pins the width of a member somebody thought about and says nothing
about one nobody did. Restated in README.md as four rules read IN ORDER, the
last of which matches everything, so no member can fail to be placed:

  1. a line or column in the resolved model, including a line count, which is
     the last line's own number — u64;
  2. an ordinal in data the producer built, which painty neither computes nor
     bounds — u64;
  3. a key into a structure painty does not own and never computes — that
     contract's width;
  4. anything else, an index or count of things in memory — usize.

Ordered rather than listed because a line count matches both 1 and 4, and the
first match has to win. Working the whole public surface through it turned up
a fifth member nobody had placed either: `Adapted::dropped_labels` and
`dropped_path_segments`, counts of labels that did not fit a caller's buffer,
which rule 4 makes `usize`.

`tests/exact_positions.rs` becomes `tests/numeric_widths.rs`, since it is no
longer only about positions, and gains two things. The ascriptions now cover
the WHOLE public numeric surface rather than the members under discussion. And
a new census asserts that `u32` appears nowhere outside rule 3's one file —
the only one of the three checks that can catch a member which does not exist
yet, because an ascription cannot.

What that still does not cover is recorded in the file rather than papered
over: a new member spelled `u64` or `usize` that should have been something
else passes every check here, and only a syn-based walk of the public surface
would catch it. That is a dependency and a machine for thirty numeric members.

+++ src/tokora.rs
@@ -113,7 +113,10 @@ impl<'a> From<::tokora::diagnostic::PathSegment<'a>> for PathSegment<'a> {
   fn from(segment: ::tokora::diagnostic::PathSegment<'a>) -> Self {
     match segment {
       ::tokora::diagnostic::PathSegment::Field(name) => Self::Field(name),
-      ::tokora::diagnostic::PathSegment::Index(index) => Self::Index(index),
+      // Widening, so nothing can be lost. painty's index is deliberately wider than tokora's:
+      // painty's model is not shaped by tokora's — that is what makes this an adapter rather than
+      // a re-export — and a result path entry has no 32-bit cap in any specification.
+      ::tokora::diagnostic::PathSegment::Index(index) => Self::Index(u64::from(index)),
     }
   }
 }
